add missing parties to routing
