+++ practice-app/frontend/src/App.js
@@ -1,9 +1,6 @@
 import React, { Component } from 'react';
-import { Route } from "react-router-dom";
 import Header from "./components/Header/Header"
 import Routes from "./components/Routes"
-import Signup from "./components/signup/SignUp"
-import Login from "./components/login/Login"
 
 import './App.css';
 
@@ -12,9 +9,6 @@ class App extends Component {
         return (
             <div className="App">
                 <Header/>
-                {
-
-                }
                 <Routes/>
             </div>
         )
