+++ practice-app/frontend/src/components/Routes.js
@@ -1,25 +1,20 @@
 import React, { Component } from "react";
-import { Router, Route, Switch, Redirect } from "react-router-dom";
+import { Router, Route, Switch } from "react-router-dom";
 import history from "./core/history"
 import SignUp from "./signup/SignUp"
-
+import TradingEq from "./t-equipments";
+import Home from "./event";
+import CurrencyConverter from "./t-equipments/currency-converter";
 class Routes extends Component {
     render() {
         return (
         <Router history={history}>
             <Switch>
-                <Route
-                    path="/signup"
-                    render={() => {
-                        return <SignUp />;
-                    }}
-                />
-                <Route
-                    path="/"
-                    render={() => {
-                        return null;
-                    }}
-                />
+                <Route path="/signup" render={() => { return <SignUp />; }} />
+                <Route path="/t-equipments/converter" render={() => { return <CurrencyConverter />; }} />
+                <Route path="/t-equipments" render={() => { return <TradingEq />; }} />
+                <Route path="/events" render={() => { return <Home />; }} />
+                <Route path="/" render={() => { return <Home />; }} />
             </Switch>
         </Router>
     )
